SCAN4NET-334 Extract ProjectKindTest
