+++ its/src/test/java/com/sonar/it/scanner/msbuild/sonarqube/ProjectKindTest.java
@@ -1,0 +1,264 @@
+/*
+ * SonarScanner for .NET
+ * Copyright (C) 2016-2025 SonarSource SA
+ * mailto:info AT sonarsource DOT com
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+package com.sonar.it.scanner.msbuild.sonarqube;
+
+import com.sonar.it.scanner.msbuild.utils.ScannerClassifier;
+import com.sonar.it.scanner.msbuild.utils.TestUtils;
+import com.sonar.orchestrator.build.BuildResult;
+import java.io.IOException;
+import java.nio.file.Path;
+import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Collectors;
+import org.apache.commons.io.FileUtils;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.junit.jupiter.api.io.TempDir;
+import org.sonarqube.ws.Ce;
+import org.sonarqube.ws.Components;
+import org.sonarqube.ws.Issues.Issue;
+import org.sonarqube.ws.client.WsClient;
+import org.sonarqube.ws.client.components.ShowRequest;
+
+import static com.sonar.it.scanner.msbuild.sonarqube.ServerTests.ORCHESTRATOR;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.tuple;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assumptions.assumeFalse;
+import static org.junit.jupiter.api.Assumptions.assumeTrue;
+
+@ExtendWith(ServerTests.class)
+class ProjectKindTest {
+  private static final String SONAR_RULES_PREFIX = "csharpsquid:";
+
+  @TempDir
+  public Path basePath;
+
+  @Test
+  void testXamlCompilation() throws IOException {
+    // We can't build with MSBuild 15
+    // error MSB4018: System.InvalidOperationException: This implementation is not part of the Windows Platform FIPS validated cryptographic algorithms.
+    // at System.Security.Cryptography.MD5CryptoServiceProvider..ctor()
+    assumeFalse(TestUtils.getMsBuildPath(ORCHESTRATOR).toString().contains("2017"));
+
+    var projectKey = "XamarinApplication";
+    BuildResult result = runAnalysis(projectKey, true);
+    assertTrue(result.isSuccess());
+
+    List<Issue> issues = TestUtils.projectIssues(ORCHESTRATOR, projectKey);
+    assertThat(issues)
+      .extracting(Issue::getRule, Issue::getComponent)
+      .containsExactlyInAnyOrder(
+        tuple(SONAR_RULES_PREFIX + "S927", "XamarinApplication:XamarinApplication.iOS/AppDelegate.cs"),
+        tuple(SONAR_RULES_PREFIX + "S927", "XamarinApplication:XamarinApplication.iOS/AppDelegate.cs"),
+        tuple(SONAR_RULES_PREFIX + "S1118", "XamarinApplication:XamarinApplication.iOS/Main.cs"),
+        tuple(SONAR_RULES_PREFIX + "S1186", "XamarinApplication:XamarinApplication.iOS/Main.cs"),
+        tuple(SONAR_RULES_PREFIX + "S1186", "XamarinApplication:XamarinApplication/App.xaml.cs"),
+        tuple(SONAR_RULES_PREFIX + "S1186", "XamarinApplication:XamarinApplication/App.xaml.cs"),
+        tuple(SONAR_RULES_PREFIX + "S1186", "XamarinApplication:XamarinApplication/App.xaml.cs"),
+        tuple(SONAR_RULES_PREFIX + "S1134", "XamarinApplication:XamarinApplication/MainPage.xaml.cs"),
+        tuple("external_roslyn:CS0618", "XamarinApplication:XamarinApplication.iOS/Main.cs"));
+
+    assertThat(TestUtils.getMeasureAsInteger("XamarinApplication", "lines", ORCHESTRATOR)).isEqualTo(149);
+    assertThat(TestUtils.getMeasureAsInteger("XamarinApplication", "ncloc", ORCHESTRATOR)).isEqualTo(93);
+    assertThat(TestUtils.getMeasureAsInteger("XamarinApplication", "files", ORCHESTRATOR)).isEqualTo(6);
+    assertThat(TestUtils.getMeasureAsInteger("XamarinApplication:XamarinApplication.iOS", "lines", ORCHESTRATOR)).isEqualTo(97);
+    assertThat(TestUtils.getMeasureAsInteger("XamarinApplication:XamarinApplication", "lines", ORCHESTRATOR)).isEqualTo(52);
+  }
+
+  @Test
+  void testRazorCompilationNet9WithoutSourceGenerators() throws IOException {
+    assumeTrue(TestUtils.getMsBuildPath(ORCHESTRATOR).toString().contains("2022")); // We can't build without MsBuild17
+    String projectName = "RazorWebApplication.net9.withoutSourceGenerators";
+    assertProjectFileContains(projectName, "<UseRazorSourceGenerator>false</UseRazorSourceGenerator>");
+    validateRazorProject(projectName);
+  }
+
+  @Test
+  void testRazorCompilationNet9WithSourceGenerators() throws IOException {
+    assumeTrue(TestUtils.getMsBuildPath(ORCHESTRATOR).toString().contains("2022")); // We can't build without MsBuild17
+    String projectName = "RazorWebApplication.net9.withSourceGenerators";
+    assertProjectFileContains(projectName, "<UseRazorSourceGenerator>true</UseRazorSourceGenerator>");
+    validateRazorProject(projectName);
+  }
+
+  @Test
+  void testCSharpAllFlat() throws IOException {
+    runAnalysis("CSharpAllFlat");
+
+    assertThat(getComponent("CSharpAllFlat:Common.cs")).isNotNull();
+  }
+
+  @Test
+  void testCSharpSharedFiles() throws IOException {
+    runAnalysis("CSharpSharedFiles");
+
+    assertThat(getComponent("CSharpSharedFiles:Common.cs"))
+      .isNotNull();
+    String class1ComponentId = TestUtils.hasModules(ORCHESTRATOR)
+      ? "CSharpSharedFiles:CSharpSharedFiles:D8FEDBA2-D056-42FB-B146-5A409727B65D:Class1.cs"
+      : "CSharpSharedFiles:ClassLib1/Class1.cs";
+    assertThat(getComponent(class1ComponentId))
+      .isNotNull();
+    String class2ComponentId = TestUtils.hasModules(ORCHESTRATOR)
+      ? "CSharpSharedFiles:CSharpSharedFiles:72CD6ED2-481A-4828-BA15-8CD5F0472A77:Class2.cs"
+      : "CSharpSharedFiles:ClassLib2/Class2.cs";
+    assertThat(getComponent(class2ComponentId))
+      .isNotNull();
+  }
+
+  @Test
+  void testCSharpSharedProjectType() throws IOException {
+    runAnalysis("CSharpSharedProjectType");
+
+    assertThat(getComponent("CSharpSharedProjectType:SharedProject/TestEventInvoke.cs"))
+      .isNotNull();
+    String programComponentId1 = TestUtils.hasModules(ORCHESTRATOR)
+      ? "CSharpSharedProjectType:CSharpSharedProjectType:36F96F66-8136-46C0-B83B-EFAE05A8FFC1:Program.cs"
+      : "CSharpSharedProjectType:ConsoleApp1/Program.cs";
+    assertThat(getComponent(programComponentId1))
+      .isNotNull();
+    String programComponentId2 = TestUtils.hasModules(ORCHESTRATOR)
+      ? "CSharpSharedProjectType:CSharpSharedProjectType:F96D8AA1-BCE1-4655-8D65-08F2A5FAC15B:Program.cs"
+      : "CSharpSharedProjectType:ConsoleApp2/Program.cs";
+    assertThat(getComponent(programComponentId2))
+      .isNotNull();
+  }
+
+  @Test
+  void testCSharpFramework48() throws IOException {
+    var folderName = "CSharp.Framework.4.8";
+    assumeFalse(TestUtils.getMsBuildPath(ORCHESTRATOR).toString().contains("2017")); // We can't run .NET Core SDK under VS 2017 CI context
+    BuildResult buildResult = runAnalysis(folderName, true);
+
+    assertUIWarnings(buildResult);
+    List<Issue> issues = TestUtils.projectIssues(ORCHESTRATOR, folderName);
+    assertThat(issues).hasSize(2)
+      .extracting(Issue::getRule, Issue::getComponent)
+      .containsExactlyInAnyOrder(
+        tuple(SONAR_RULES_PREFIX + "S1134", folderName + ":Main/Common.cs"),
+        tuple(SONAR_RULES_PREFIX + "S2699", folderName + ":UTs/CommonTest.cs"));
+  }
+
+  @Test
+  void testCSharpSdk8() throws IOException {
+    validateCSharpSdk("CSharp.SDK.8");
+  }
+
+  @Test
+  void testScannerNet8NoAnalysisWarnings() throws IOException {
+    // dotnet sdk tests should run only on VS 2022
+    assumeTrue(TestUtils.getMsBuildPath(ORCHESTRATOR).toString().contains("2022"));
+
+    BuildResult buildResult = runAnalysis("CSharp.SDK.8");
+
+    assertThat(buildResult.getLogs()).doesNotContain("Failed to parse properties from the environment variable 'SONARQUBE_SCANNER_PARAMS'");
+    assertUIWarnings(buildResult);
+  }
+
+  @Test
+  void testCSharpSdkLatest() throws IOException {
+    validateCSharpSdk("CSharp.SDK.Latest");
+  }
+
+  private void validateCSharpSdk(String folderName) throws IOException {
+    // dotnet sdk tests should run only on VS 2022
+    assumeTrue(TestUtils.getMsBuildPath(ORCHESTRATOR).toString().contains("2022"));
+
+    runAnalysis(folderName);
+
+    List<Issue> issues = TestUtils.projectIssues(ORCHESTRATOR, folderName);
+
+    assertThat(issues).hasSize(2)
+      .extracting(Issue::getRule, Issue::getComponent)
+      .containsExactlyInAnyOrder(
+        tuple(SONAR_RULES_PREFIX + "S1134", folderName + ":Main/Common.cs"),
+        tuple(SONAR_RULES_PREFIX + "S2699", folderName + ":UTs/CommonTest.cs"));
+    // The AspNetCoreMvc/Views/Home/Index.cshtml contains an external CS0219 issue
+    // which is currently not imported due to the fact that the generated code Index.cshtml.g.cs is in the object folder.
+  }
+
+  private void assertUIWarnings(BuildResult buildResult) {
+    // AnalysisWarningsSensor was implemented starting from analyzer version 8.39.0.47922 (https://github.com/SonarSource/sonar-dotnet-enterprise/commit/39baabb01799aa1945ac5c80d150f173e6ada45f)
+    // So it's available from SQ 9.9 onwards
+    if (ORCHESTRATOR.getServer().version().isGreaterThanOrEquals(9, 9)) {
+      var warnings = TestUtils.getAnalysisWarningsTask(ORCHESTRATOR, buildResult);
+      assertThat(warnings.getStatus()).isEqualTo(Ce.TaskStatus.SUCCESS);
+      var warningsList = warnings.getWarningsList();
+      assertThat(warningsList.stream().anyMatch(
+        // The warning is appended to the timestamp, we want to assert only the message
+        x -> x.endsWith("Multi-Language analysis is enabled. If this was not intended and you have issues such as hitting your LOC limit or analyzing unwanted files, please set " +
+          "\"/d:sonar.scanner.scanAll=false\" in the begin step.")
+      )).isTrue();
+      assertThat(warningsList.size()).isEqualTo(1);
+    }
+  }
+
+  private void assertProjectFileContains(String projectName, String textToLookFor) throws IOException {
+    Path projectPath = TestUtils.projectDir(basePath, projectName);
+    Path csProjPath = projectPath.resolve("RazorWebApplication\\RazorWebApplication.csproj");
+    String str = FileUtils.readFileToString(csProjPath.toFile(), "utf-8");
+    assertThat(str.indexOf(textToLookFor)).isPositive();
+  }
+
+  private BuildResult runAnalysis(String folderName) throws IOException {
+    return runAnalysis(folderName, false);
+  }
+
+  private BuildResult runAnalysis(String folderName, Boolean useNuGet) throws IOException {
+    Path projectDir = TestUtils.projectDir(basePath, folderName);
+    return TestUtils.runAnalysis(projectDir, folderName, useNuGet);
+  }
+
+  private void validateRazorProject(String projectKey) throws IOException {
+    ORCHESTRATOR.getServer().provisionProject(projectKey, projectKey);
+
+    if (TestUtils.getMsBuildPath(ORCHESTRATOR).toString().contains("2017")) {
+      return; // We can't build razor under VS 2017 CI context
+    }
+
+    Path projectDir = TestUtils.projectDir(basePath, projectKey);
+    String token = TestUtils.getNewToken(ORCHESTRATOR);
+
+    TestUtils.newScannerBegin(ORCHESTRATOR, projectKey, projectDir, token, ScannerClassifier.NET_FRAMEWORK).execute(ORCHESTRATOR);
+    TestUtils.runNuGet(ORCHESTRATOR, projectDir, false, "restore");
+    TestUtils.runDotnetCommand(projectDir, "build", "--no-incremental");
+    BuildResult result = TestUtils.executeEndStepAndDumpResults(ORCHESTRATOR, projectDir, projectKey, token);
+
+    assertTrue(result.isSuccess());
+
+    List<Issue> issues = TestUtils.projectIssues(ORCHESTRATOR, projectKey);
+    List<String> ruleKeys = issues.stream().map(Issue::getRule).collect(Collectors.toList());
+
+    assertThat(ruleKeys).containsAll(Arrays.asList(SONAR_RULES_PREFIX + "S1118", SONAR_RULES_PREFIX + "S1186"));
+
+    assertThat(TestUtils.getMeasureAsInteger(projectKey, "lines", ORCHESTRATOR)).isEqualTo(49);
+    assertThat(TestUtils.getMeasureAsInteger(projectKey, "ncloc", ORCHESTRATOR)).isEqualTo(39);
+    assertThat(TestUtils.getMeasureAsInteger(projectKey, "files", ORCHESTRATOR)).isEqualTo(2);
+  }
+
+  private static Components.Component getComponent(String componentKey) {
+    return newWsClient().components().show(new ShowRequest().setComponent(componentKey)).getComponent();
+  }
+
+  private static WsClient newWsClient() {
+    return TestUtils.newWsClient(ORCHESTRATOR);
+  }
+}
